qa_app/requirements.txt
flask==2.0.1
flask-sqlalchemy==3.0.2
openai==1.0.0
python-dotenv==0.19.0 